test: StoreReturn with EAGER
