+++ test-suite/src/test/java/io/micronaut/microstream/docs/CrmCustomerService.java
@@ -1,31 +1,11 @@
 package io.micronaut.microstream.docs;
 
 import io.micronaut.core.annotation.NonNull;
-import io.micronaut.microstream.annotations.Store;
-import io.micronaut.microstream.annotations.StoringStrategy;
-import jakarta.inject.Singleton;
-import one.microstream.storage.types.StorageManager;
 
 import javax.validation.Valid;
 import javax.validation.constraints.NotNull;
 
-@Singleton
-public class CrmCustomerService {
-
-    private final StorageManager storageManager;
-
-    public CrmCustomerService(StorageManager storageManager) {
-        this.storageManager = storageManager;
-    }
-
-    @Store(result = true, strategy = StoringStrategy.EAGER)
+public interface CrmCustomerService {
     @NonNull
-    public Customers save(@NonNull @NotNull @Valid Customer customer) {
-        data().getCustomers().getCustomersById().put(customer.getId(), customer);
-        return data().getCustomers();
-    }
-
-    private CRM data() {
-        return (CRM) storageManager.root();
-    }
+    Customers save(@NonNull @NotNull @Valid Customer customer);
 }
