+++ test-suite/src/test/java/io/micronaut/microstream/docs/EagerTest.java
@@ -3,7 +3,8 @@ package io.micronaut.microstream.docs;
 import io.micronaut.context.ApplicationContext;
 import io.micronaut.core.util.CollectionUtils;
 import one.microstream.storage.types.StorageManager;
-import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 
 import java.util.Map;
 import java.util.UUID;
@@ -13,13 +14,17 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 
 class EagerTest {
 
-    @Test
-    void testEagerStrategyWithReturn() {
+    @ParameterizedTest
+    @ValueSource(strings = {
+        "store",
+        "store-return"
+    })
+    void testEagerStrategyWithReturn(String serviceImplementation) {
         String storageDirectory = "build/microstream-" + UUID.randomUUID();
         Map<String, Object> properties = CollectionUtils.mapOf(
             "microstream.storage.main.storage-directory", storageDirectory,
-            "microstream.storage.main.root-class", "io.micronaut.microstream.docs.CRM"
-        );
+            "microstream.storage.main.root-class", "io.micronaut.microstream.docs.CRM",
+            "spec.service", serviceImplementation);
         ApplicationContext ctx = ApplicationContext.run(properties);
         CRM data = (CRM) ctx.getBean(StorageManager.class).root();
         assertTrue(data.getCustomers().getCustomersById().isEmpty());
